+++ src/menu-entry/menuDialog_ui_derived.h
@@ -3,16 +3,28 @@
 
 #include "menuDialog_ui_base.h"
 
+enum class DialogAction
+{
+  None = 0,
+  UpdateOwnShip = 1,
+  UpdateAisTarget = 2,
+  AddAisTarget = 3,
+  RemoveLastAisTarget = 4,
+  ClearAisTargets = 5
+};
 
 class DialogMenuEntry : public DialogMenuEntryBase
 {
   public:
     DialogMenuEntry( wxWindow* parent );
-    int m_updateVesselPositionChoice = 0;
+    DialogAction m_action = DialogAction::None;
 
   protected:
     void OnButtonClick_UpdateOwnShipPosition( wxCommandEvent& event ) override;
     void OnButtonClick_UpdateAisTargetPosition( wxCommandEvent& event ) override;
+    void OnButtonClick_AddAisTarget( wxCommandEvent& event ) override;
+    void OnButtonClick_RemoveLastAisTargets( wxCommandEvent& event ) override;
+    void OnButtonClick_ClearAisTargets( wxCommandEvent& event) override;
 };
 
 #endif // __menuDialog_ui_derived__